newline by default, lc to continue line
after experimentation, it seems nicer to assume each element is
its own line and using the occasional `lc` (line-continue)
marker to have one element directly follow the former, than it
is to use `fl` (fresh line) every time the output should start
on a new line.

It is expressionally equivalent, but a question of pain allocation,
and I think this new approach is less overhead.