block: >> and <<end>> tags must be aligned with <<crowbar
`>>` can be interpreted as a bit shift operator in many languages.
To work around this and other potential issues, we now require that
the magic tokens (`<<crowbar`, `>>` and `<<end>>`) are aligned to
the same column.
In any other instance, they would be considered normal content.
This both makes accidental matching very unlikely AND enforces
a nicer formatting of blocks (at the cost of rigidity).